Ah I see the source of my misery.
So it turns out that go's append function is very weird. If the slice
has enough memory, it will modify the slice in place. But if the slice
does not, it will create a copy of the slice. Meaning when i was
building my backtracking list, i could have modified an already built
slice previously (Mutation is deadly...), causing me to purchase more
than I had in budget even though at face value my logic prevented it.